refactor(code-ops): split AST file operations
